Lint entire project
