+++ .swiftlint.yml
@@ -1,0 +1,4 @@
+opt_in_rules:
+  - empty_count
+  - vertical_whitespace
+line_length: 220
