+++ Sources/Deferred/Executor.swift
@@ -142,4 +142,3 @@ extension RunLoop: Executor {
         getCFRunLoop().submit(body)
     }
 }
-
