+++ Tests/DeferredTests/DeferredTests.swift
@@ -17,8 +17,10 @@ import Dispatch
     import Glibc
 #endif
 
-class DeferredTests: XCTestCase {
+// swiftlint:disable type_body_length
+// We wanna test things!
 
+class DeferredTests: XCTestCase {
     static var allTests: [(String, (DeferredTests) -> () throws -> Void)] {
         let universalTests: [(String, (DeferredTests) -> () throws -> Void)] = [
             ("testDestroyedWithoutBeingFilled", testDestroyedWithoutBeingFilled),
@@ -35,10 +37,6 @@ class DeferredTests: XCTestCase {
             ("testUponCalledWhenFilled", testUponCalledWhenFilled),
             ("testUponMainQueueCalledWhenFilled", testUponMainQueueCalledWhenFilled),
             ("testConcurrentUpon", testConcurrentUpon),
-            ("testAnd", testAnd),
-            ("testAllFilled", testAllFilled),
-            ("testAllFilledEmptyCollection", testAllFilledEmptyCollection),
-            ("testFirstFilled", testFirstFilled),
             ("testAllCopiesOfADeferredValueRepresentTheSameDeferredValue", testAllCopiesOfADeferredValueRepresentTheSameDeferredValue),
             ("testDeferredOptionalBehavesCorrectly", testDeferredOptionalBehavesCorrectly),
             ("testIsFilledCanBeCalledMultipleTimesNotFilled", testIsFilledCanBeCalledMultipleTimesNotFilled),
@@ -240,89 +238,6 @@ class DeferredTests: XCTestCase {
         waitForExpectations()
     }
 
-    func testAnd() {
-        let d1 = Deferred<Int>()
-        let d2 = Deferred<String>()
-        let both = d1.and(d2)
-
-        XCTAssertFalse(both.isFilled)
-
-        d1.fill(with: 1)
-        XCTAssertFalse(both.isFilled)
-        d2.fill(with: "foo")
-
-        let expectation = self.expectation(description: "paired deferred should be filled")
-        both.upon { _ in
-            XCTAssert(d1.isFilled)
-            XCTAssert(d2.isFilled)
-            XCTAssertEqual(both.value.0, 1)
-            XCTAssertEqual(both.value.1, "foo")
-            expectation.fulfill()
-        }
-
-        waitForExpectations()
-    }
-
-    func testAllFilled() {
-        var d = [Deferred<Int>]()
-
-        for _ in 0 ..< 10 {
-            d.append(Deferred())
-        }
-
-        let w = d.allFilled()
-        let outerExpectation = expectation(description: "all results filled in")
-        let innerExpectation = expectation(description: "paired deferred should be filled")
-
-        // skip first
-        for i in 1 ..< d.count {
-            d[i].fill(with: i)
-        }
-
-        afterDelay {
-            XCTAssertFalse(w.isFilled) // unfilled because d[0] is still unfilled
-            d[0].fill(with: 0)
-
-            self.afterDelay {
-                XCTAssertTrue(w.value == [Int](0 ..< d.count))
-                innerExpectation.fulfill()
-            }
-            outerExpectation.fulfill()
-        }
-
-        waitForExpectations()
-    }
-
-    func testAllFilledEmptyCollection() {
-        let d = EmptyCollection<Deferred<Int>>().allFilled()
-        XCTAssert(d.isFilled)
-    }
-
-    func testFirstFilled() {
-        let d = (0 ..< 10).map { _ in Deferred<Int>() }
-        let w = d.firstFilled()
-
-        d[3].fill(with: 3)
-
-        let outerExpectation = expectation(description: "any is filled")
-        let innerExpectation = expectation(description: "any is not changed")
-
-        afterDelay {
-            XCTAssertEqual(w.value, 3)
-
-            d[4].fill(with: 4)
-
-            self.afterDelay {
-                XCTAssertEqual(w.value, 3)
-                innerExpectation.fulfill()
-            }
-
-            outerExpectation.fulfill()
-        }
-
-        waitForExpectations()
-    }
-
     /// Deferred values behave as values: All copies reflect the same value.
     /// The wrinkle of course is that the value might not be observable till a later
     /// date.
@@ -448,5 +363,6 @@ class DeferredTests: XCTestCase {
         XCTAssertNotNil(deferred.wait(until: .now()))
         XCTAssertNotNil(deferred.waitShort())  // pass
     }
-
 }
+
+// swiftlint:enable type_body_length
