+++ Tests/LinuxMain.swift
@@ -15,6 +15,7 @@ XCTMain([
     testCase(DeferredTests.allTests),
     testCase(ExistentialFutureTests.allTests),
     testCase(FutureCustomExecutorTests.allTests),
+    testCase(FutureTests.allTests),
     testCase(IgnoringFutureTests.allTests),
     testCase(LockingTests.allTests),
     testCase(ProtectedTests.allTests),
