+++ Tests/TaskTests/TaskGroupTests.swift
@@ -19,8 +19,7 @@ import Deferred
 #endif
 
 class TaskGroupTests: XCTestCase {
-
-    static var allTests : [(String, (TaskGroupTests) -> () throws -> Void)] {
+    static var allTests: [(String, (TaskGroupTests) -> () throws -> Void)] {
         return [
             ("testThatAllCompleteTaskWaitsForAllAccumulatedTasks", testThatAllCompleteTaskWaitsForAllAccumulatedTasks),
         ]
